safari-534.53-branch/LayoutTests/http/tests/media/video-play-progress-expected.txt

Test that at least one progress event is fired after starting to load the video.

RUN(video.play())
EVENT(loadstart)
EVENT(progress)
END OF TEST
